QLDRL/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cloudinary==1.40.0
Django==4.2.13
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
PyMySQL==1.1.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
